+++ src/basket/styles/basket-mini-view.module.scss
